Version bump to 2.0.0-beta.24.volders
We need to enable the source map generation, but doing so we had this
issue (https://github.com/ember-intl/ember-intl/issues/204).

To solve it, we need to update the ember-intl to at least version 2.0.1
to get this tiny fix:

  https://github.com/ember-intl/ember-intl/pull/254/commits/1ce1ecbf855745ddd2e3b95f06b01e037fa6bf79

The problem is that when updating it, it updates a lot of other packages
and we start getting new errors.

So, to fix the issue without the need to update other packages right
now, we created this version cherry-picking the fix commit on top of the
version we are currently using (2.0.0-beta.24).